pyproject.toml
[tool.poetry]
name = "qac-knowledge-graph"
version = "0.1.0"
description = ""
authors = ["Tulio Ribeiro <tulio@britoribeiro.net>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12"
faiss-cpu = "^1.8.0.post1"
nltk = "^3.9.1"
pypdf2 = "^3.0.1"
fitz = "^0.0.1.dev2"
beautifulsoup4 = "^4.12.3"
python-docx = "^1.1.2"
pymupdf = "^1.24.10"
spacy = "^3.7.6"
python-dotenv = "^1.0.1"
pillow = "^10.4.0"
pytesseract = "^0.3.13"
tenacity = "^9.0.0"
gpt4all = "^2.8.2"
llama-toolchain = "^0.0.17"


[tool.poetry.group.dev.dependencies]
notebook = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
